refactor(main): register IPC handlers per domain outside the entry
