+++ src/main/renderer-ipc.test.ts
@@ -1,5 +1,6 @@
 // @vitest-environment node
 
+import { IPC_CHANNELS } from "@openbot/contracts/ipc";
 import { describe, expect, it, vi } from "vitest";
 import { type RendererIpcWindow, sendToRenderer } from "./renderer-ipc";
 
@@ -34,8 +35,8 @@ describe("renderer IPC", () => {
   it("sends to a live renderer frame", () => {
     const window = rendererWindow();
 
-    expect(sendToRenderer(window, "status", { ready: true })).toBe(true);
-    expect(window.webContents.send).toHaveBeenCalledWith("status", { ready: true });
+    expect(sendToRenderer(window, IPC_CHANNELS.updateEvent, { ready: true })).toBe(true);
+    expect(window.webContents.send).toHaveBeenCalledWith(IPC_CHANNELS.updateEvent, { ready: true });
   });
 
   it.each([
@@ -47,7 +48,7 @@ describe("renderer IPC", () => {
   ])("does not send during a %s", (_name, overrides) => {
     const window = rendererWindow(overrides);
 
-    expect(sendToRenderer(window, "status")).toBe(false);
+    expect(sendToRenderer(window, IPC_CHANNELS.updateEvent)).toBe(false);
     expect(window.webContents.send).not.toHaveBeenCalled();
   });
 
@@ -56,12 +57,12 @@ describe("renderer IPC", () => {
     Object.assign(new Error("write EPIPE"), { code: "EPIPE" }),
     new Error("Render frame was disposed before WebFrameMain could be accessed"),
   ])("contains an unavailable-renderer race without crashing the main process", (error) => {
-    expect(sendToRenderer(rendererWindow({ sendError: error }), "status")).toBe(false);
+    expect(sendToRenderer(rendererWindow({ sendError: error }), IPC_CHANNELS.updateEvent)).toBe(false);
   });
 
   it("does not hide unrelated IPC errors", () => {
     const error = new Error("An object could not be cloned");
 
-    expect(() => sendToRenderer(rendererWindow({ sendError: error }), "status")).toThrow(error);
+    expect(() => sendToRenderer(rendererWindow({ sendError: error }), IPC_CHANNELS.updateEvent)).toThrow(error);
   });
 });
